+++ src/Pim/Bundle/CatalogBundle/Doctrine/ORM/Repository/GroupTypeRepository.php
@@ -30,29 +30,6 @@ class GroupTypeRepository extends EntityRepository implements GroupTypeRepositor
     /**
      * {@inheritdoc}
      */
-    public function createDatagridQueryBuilder()
-    {
-        $rootAlias = 'g';
-        $qb = $this->createQueryBuilder($rootAlias);
-
-        $labelExpr = sprintf(
-            "(CASE WHEN translation.label IS NULL THEN %s.code ELSE translation.label END)",
-            $rootAlias
-        );
-
-        $qb
-            ->addSelect($rootAlias)
-            ->addSelect(sprintf("%s AS label", $labelExpr));
-
-        $qb
-            ->leftJoin($rootAlias .'.translations', 'translation', 'WITH', 'translation.locale = :localeCode');
-
-        return $qb;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function getTypeByGroup($code)
     {
         try {
