+++ src/Pim/Bundle/DataGridBundle/Datasource/Datasource.php
@@ -8,7 +8,7 @@ use Doctrine\ORM\QueryBuilder;
 use Oro\Bundle\DataGridBundle\Datagrid\DatagridInterface;
 use Pim\Bundle\DataGridBundle\Datagrid\Configuration\Product\ContextConfigurator;
 use Pim\Bundle\DataGridBundle\Datasource\ResultRecord\HydratorInterface;
-use Pim\Component\Catalog\Repository\MassActionRepositoryInterface;
+use Pim\Bundle\DataGridBundle\Doctrine\ORM\Repository\MassActionRepositoryInterface;
 
 /**
  * Pim agnostic datasource
@@ -16,6 +16,8 @@ use Pim\Component\Catalog\Repository\MassActionRepositoryInterface;
  * @author    Julien Janvier <julien.janvier@akeneo.com>
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ *
+ * @deprecated you should use RepositoryDatasource
  */
 class Datasource implements DatasourceInterface, ParameterizableInterface
 {
@@ -91,24 +93,6 @@ class Datasource implements DatasourceInterface, ParameterizableInterface
     public function getQueryBuilder()
     {
         return $this->qb;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setQueryBuilder($qb)
-    {
-        $this->qb = $qb;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getObjectManager()
-    {
-        return $this->om;
     }
 
     /**
