+++ src/Pim/Bundle/DataGridBundle/Extension/MassAction/MassActionDispatcher.php
@@ -158,10 +158,8 @@ class MassActionDispatcher
         $this->requestParams->set(FilterExtension::FILTER_ROOT_PARAM, $filters);
 
         $qb = $datagrid->getAcceptedDatasource()->getQueryBuilder();
-
         if (self::FAMILY_GRID_NAME === $datagridName) {
             $qbLocaleParameter = $qb->getParameter('localeCode');
-
             if (null !== $qbLocaleParameter && null === $qbLocaleParameter->getValue()) {
                 $qb->setParameter('localeCode', $request->query->get('dataLocale'));
             }
