+++ src/Pim/Bundle/DataGridBundle/Resources/config/data_sources.yml
@@ -1,6 +1,8 @@
 parameters:
     pim_datagrid.datasource.default.class: Pim\Bundle\DataGridBundle\Datasource\Datasource
     pim_datagrid.datasource.product.class: Pim\Bundle\DataGridBundle\Datasource\ProductDatasource
+    pim_datagrid.datasource.family.class: Pim\Bundle\DataGridBundle\Datasource\FamilyDatasource
+    pim_datagrid.datasource.repository.class: Pim\Bundle\DataGridBundle\Datasource\RepositoryDatasource
     pim_datagrid.datasource.support_resolver.class: Pim\Bundle\DataGridBundle\Datasource\DatasourceSupportResolver
     pim_datagrid.datasource.adapter_resolver.class: Pim\Bundle\DataGridBundle\Datasource\DatasourceAdapterResolver
 
@@ -12,6 +14,87 @@ services:
             - '@pim_datagrid.datasource.result_record.hydrator.default'
         tags:
             - { name: oro_datagrid.datasource, type: pim_datasource_default }
+
+    pim_datagrid.datasource.attribute:
+        class: '%pim_datagrid.datasource.repository.class%'
+        arguments:
+            - '@pim_enrich.repository.attribute'
+            - '@pim_datagrid.datasource.result_record.hydrator.default'
+        tags:
+            - { name: oro_datagrid.datasource, type: pim_datasource_attribute }
+
+    pim_datagrid.datasource.family:
+        class: '%pim_datagrid.datasource.family.class%'
+        arguments:
+            - '@pim_enrich.repository.family'
+            - '@pim_enrich.repository.family'
+            - '@pim_datagrid.datasource.result_record.hydrator.default'
+        tags:
+            - { name: oro_datagrid.datasource, type: pim_datasource_family }
+
+    pim_datagrid.datasource.group_type:
+        class: '%pim_datagrid.datasource.repository.class%'
+        arguments:
+            - '@pim_enrich.repository.group_type'
+            - '@pim_datagrid.datasource.result_record.hydrator.default'
+        tags:
+            - { name: oro_datagrid.datasource, type: pim_datasource_group_type }
+
+    pim_datagrid.datasource.association_type:
+        class: '%pim_datagrid.datasource.repository.class%'
+        arguments:
+            - '@pim_enrich.repository.association_type'
+            - '@pim_datagrid.datasource.result_record.hydrator.default'
+        tags:
+            - { name: oro_datagrid.datasource, type: pim_datasource_association_type }
+
+    pim_datagrid.datasource.group:
+        class: '%pim_datagrid.datasource.repository.class%'
+        arguments:
+            - '@pim_enrich.repository.group'
+            - '@pim_datagrid.datasource.result_record.hydrator.default'
+        tags:
+            - { name: oro_datagrid.datasource, type: pim_datasource_group }
+
+    pim_datagrid.datasource.locale:
+        class: '%pim_datagrid.datasource.repository.class%'
+        arguments:
+            - '@pim_enrich.repository.locale'
+            - '@pim_datagrid.datasource.result_record.hydrator.default'
+        tags:
+            - { name: oro_datagrid.datasource, type: pim_datasource_locale }
+
+    pim_datagrid.datasource.channel:
+        class: '%pim_datagrid.datasource.repository.class%'
+        arguments:
+            - '@pim_enrich.repository.channel'
+            - '@pim_datagrid.datasource.result_record.hydrator.default'
+        tags:
+            - { name: oro_datagrid.datasource, type: pim_datasource_channel }
+
+    pim_datagrid.datasource.currency:
+        class: '%pim_datagrid.datasource.repository.class%'
+        arguments:
+            - '@pim_enrich.repository.currency'
+            - '@pim_datagrid.datasource.result_record.hydrator.default'
+        tags:
+            - { name: oro_datagrid.datasource, type: pim_datasource_currency }
+
+    pim_datagrid.datasource.job_instance:
+        class: '%pim_datagrid.datasource.repository.class%'
+        arguments:
+            - '@pim_enrich.repository.job_instance'
+            - '@pim_datagrid.datasource.result_record.hydrator.default'
+        tags:
+            - { name: oro_datagrid.datasource, type: pim_datasource_job_instance }
+
+    pim_datagrid.datasource.job_execution:
+        class: '%pim_datagrid.datasource.repository.class%'
+        arguments:
+            - '@pim_enrich.repository.job_execution'
+            - '@pim_datagrid.datasource.result_record.hydrator.default'
+        tags:
+            - { name: oro_datagrid.datasource, type: pim_datasource_job_execution }
 
     pim_datagrid.datasource.smart:
         class: '%pim_datagrid.datasource.default.class%'
