+++ src/Pim/Bundle/EnrichBundle/Doctrine/ORM/Repository/AssociationTypeRepository.php
@@ -1,0 +1,50 @@
+<?php
+
+namespace Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository;
+
+use Doctrine\ORM\EntityManager;
+use Doctrine\ORM\EntityRepository;
+use Pim\Bundle\DataGridBundle\Doctrine\ORM\Repository\DatagridRepositoryInterface;
+
+/**
+ * Association type repository
+ *
+ * @author    Gildas Quemener <gildas@akeneo.com>
+ * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+class AssociationTypeRepository extends EntityRepository implements DatagridRepositoryInterface
+{
+    /**
+     * @param EntityManager $em
+     * @param string        $class
+     */
+    public function __construct(EntityManager $em, $class)
+    {
+        parent::__construct($em, $em->getClassMetadata($class));
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function createDatagridQueryBuilder()
+    {
+        $qb = $this->createQueryBuilder('a');
+        $rootAlias = $qb->getRootAlias();
+
+        $labelExpr = sprintf(
+            "(CASE WHEN translation.label IS NULL THEN %s.code ELSE translation.label END)",
+            $rootAlias
+        );
+
+        $qb
+            ->addSelect($rootAlias)
+            ->addSelect(sprintf("%s AS label", $labelExpr))
+            ->addSelect('translation.label');
+
+        $qb
+            ->leftJoin($rootAlias .'.translations', 'translation', 'WITH', 'translation.locale = :localeCode');
+
+        return $qb;
+    }
+}
