+++ src/Pim/Bundle/EnrichBundle/Doctrine/ORM/Repository/AttributeRepository.php
@@ -4,6 +4,7 @@ namespace Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository;
 
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\EntityRepository;
+use Pim\Bundle\DataGridBundle\Doctrine\ORM\Repository\DatagridRepositoryInterface;
 use Pim\Bundle\UserBundle\Context\UserContext;
 use Pim\Component\Enrich\Provider\TranslatedLabelsProviderInterface;
 
@@ -12,7 +13,9 @@ use Pim\Component\Enrich\Provider\TranslatedLabelsProviderInterface;
  * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class AttributeRepository extends EntityRepository implements TranslatedLabelsProviderInterface
+class AttributeRepository extends EntityRepository implements
+    TranslatedLabelsProviderInterface,
+    DatagridRepositoryInterface
 {
     /** @var UserContext */
     protected $userContext;
@@ -62,5 +65,34 @@ class AttributeRepository extends EntityRepository implements TranslatedLabelsPr
         }
 
         return $choices;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function createDatagridQueryBuilder()
+    {
+        $qb = $this->createQueryBuilder('a');
+        $rootAlias = $qb->getRootAlias();
+
+        $labelExpr = sprintf(
+            '(CASE WHEN translation.label IS NULL THEN %s.code ELSE translation.label END)',
+            $rootAlias
+        );
+        $groupExpr = '(CASE WHEN gt.label IS NULL THEN attributeGroup.code ELSE gt.label END)';
+
+        $qb
+            ->addSelect($rootAlias)
+            ->addSelect(sprintf("%s AS label", $labelExpr))
+            ->addSelect(sprintf("%s AS groupLabel", $groupExpr))
+            ->addSelect('translation.label')
+            ->addSelect('attributeGroup.code');
+
+        $qb
+            ->leftJoin($rootAlias .'.translations', 'translation', 'WITH', 'translation.locale = :localeCode')
+            ->leftJoin($rootAlias .'.group', 'attributeGroup')
+            ->leftJoin('attributeGroup.translations', 'gt', 'WITH', 'gt.locale = :localeCode');
+
+        return $qb;
     }
 }
