+++ src/Pim/Bundle/EnrichBundle/Doctrine/ORM/Repository/FamilyRepository.php
@@ -4,6 +4,9 @@ namespace Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository;
 
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\EntityRepository;
+use Doctrine\ORM\QueryBuilder;
+use Pim\Bundle\DataGridBundle\Doctrine\ORM\Repository\DatagridRepositoryInterface;
+use Pim\Bundle\DataGridBundle\Doctrine\ORM\Repository\MassActionRepositoryInterface;
 use Pim\Bundle\UserBundle\Context\UserContext;
 use Pim\Component\Enrich\Provider\TranslatedLabelsProviderInterface;
 
@@ -12,7 +15,10 @@ use Pim\Component\Enrich\Provider\TranslatedLabelsProviderInterface;
  * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class FamilyRepository extends EntityRepository implements TranslatedLabelsProviderInterface
+class FamilyRepository extends EntityRepository implements
+    TranslatedLabelsProviderInterface,
+    DatagridRepositoryInterface,
+    MassActionRepositoryInterface
 {
     /** @var UserContext */
     protected $userContext;
@@ -48,5 +54,67 @@ class FamilyRepository extends EntityRepository implements TranslatedLabelsProvi
         }
 
         return $choices;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function createDatagridQueryBuilder()
+    {
+        $qb = $this->createQueryBuilder('f');
+        $rootAlias = $qb->getRootAlias();
+
+        $labelExpr = sprintf(
+            '(CASE WHEN translation.label IS NULL THEN %s.code ELSE translation.label END)',
+            $rootAlias
+        );
+
+        $qb
+            ->addSelect($rootAlias)
+            ->addSelect(sprintf('%s AS familyLabel', $labelExpr))
+            ->addSelect('translation.label');
+
+        $qb
+            ->leftJoin($rootAlias . '.translations', 'translation', 'WITH', 'translation.locale = :localeCode')
+            ->leftJoin($rootAlias . '.attributeAsLabel', 'a');
+
+        return $qb;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function applyMassActionParameters($qb, $inset, array $values)
+    {
+        if ($values) {
+            $rootAlias = $qb->getRootAlias();
+            $valueWhereCondition =
+                $inset
+                    ? $qb->expr()->in($rootAlias, $values)
+                    : $qb->expr()->notIn($rootAlias, $values);
+            $qb->andWhere($valueWhereCondition);
+        }
+
+        if (null !== $qb->getDQLPart('where')) {
+            $whereParts = $qb->getDQLPart('where')->getParts();
+            $qb->resetDQLPart('where');
+
+            foreach ($whereParts as $part) {
+                if (!is_string($part) || !strpos($part, 'entityIds')) {
+                    $qb->andWhere($part);
+                }
+            }
+        }
+
+        $qb->setParameters(
+            $qb->getParameters()->filter(
+                function ($parameter) {
+                    return $parameter->getName() !== 'entityIds';
+                }
+            )
+        );
+
+        // remove limit of the query
+        $qb->setMaxResults(null);
     }
 }
