+++ src/Pim/Bundle/EnrichBundle/Doctrine/ORM/Repository/LocaleRepository.php
@@ -1,0 +1,39 @@
+<?php
+
+namespace Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository;
+
+use Doctrine\ORM\EntityManager;
+use Doctrine\ORM\EntityRepository;
+use Pim\Bundle\DataGridBundle\Doctrine\ORM\Repository\DatagridRepositoryInterface;
+
+/**
+ * Locale repository
+ *
+ * @author    Romain Monceau <romain@akeneo.com>
+ * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+class LocaleRepository extends EntityRepository implements DatagridRepositoryInterface
+{
+    /**
+     * @param EntityManager $em
+     * @param string        $class
+     */
+    public function __construct(EntityManager $em, $class)
+    {
+        parent::__construct($em, $em->getClassMetadata($class));
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function createDatagridQueryBuilder()
+    {
+        $qb = $this->createQueryBuilder('l');
+        $rootAlias = $qb->getRootAlias();
+
+        $qb->addSelect($rootAlias);
+
+        return $qb;
+    }
+}
