+++ src/Pim/Bundle/EnrichBundle/Resources/config/datagrid/attribute.yml
@@ -1,12 +1,16 @@
 datagrid:
     attribute-grid:
-        extend: translatable-entity-grid
         options:
+            locale_parameter: localeCode
             entityHint: attribute
         source:
+            type: pim_datasource_attribute
             acl_resource: pim_enrich_attribute_index
-            entity: '%pim_catalog.entity.attribute.class%'
         columns:
+            code:
+                label: Code
+            label:
+                label: Label
             attributeType:
                 label: Type
                 type: translateable
@@ -20,10 +24,29 @@ datagrid:
                 label:     Group
                 data_name: groupLabel
         properties:
+            id: ~
             edit_link:
+                type: url
                 route: pim_enrich_attribute_edit
+                params:
+                    - id
             delete_link:
+                type: url
                 route: pim_enrich_attribute_remove
+                params:
+                    - id
+        actions:
+            edit:
+                type:      navigate
+                label:     Edit
+                icon:      pencil
+                link:      edit_link
+                rowAction: true
+            delete:
+                type:  delete
+                label: Delete
+                icon:  trash
+                link:  delete_link
         sorters:
             columns:
                 code:
@@ -34,10 +57,18 @@ datagrid:
                     data_name: a.localizable
                 group:
                     data_name: groupLabel
+                label:
+                    data_name: label
         filters:
             columns:
                 code:
+                    type:  string
+                    label: Code
                     data_name: a.code
+                label:
+                    type: string
+                    label: Label
+                    data_name: translation.label
                 attributeType:
                     type:      choice
                     data_name: a.attributeType
