+++ src/Pim/Bundle/EnrichBundle/Resources/config/datagrid/channel.yml
@@ -5,9 +5,7 @@ datagrid:
             locale_parameter: localeCode
         source:
             acl_resource: pim_enrich_channel_index
-            type: pim_datasource_default
-            entity: '%pim_catalog.entity.channel.class%'
-            repository_method: createDatagridQueryBuilder
+            type: pim_datasource_channel
         columns:
             code:
                 label: Code
