+++ src/Pim/Bundle/EnrichBundle/Resources/config/datagrid/currency.yml
@@ -4,9 +4,7 @@ datagrid:
             entityHint: currency
         source:
             acl_resource: pim_enrich_currency_index
-            type: pim_datasource_default
-            entity: '%pim_catalog.entity.currency.class%'
-            repository_method: createDatagridQueryBuilder
+            type: pim_datasource_currency
         columns:
             code:
                 label: Code
