+++ src/Pim/Bundle/EnrichBundle/Resources/config/datagrid/group.yml
@@ -4,9 +4,7 @@ datagrid:
             entityHint: group
         source:
             acl_resource: pim_enrich_group_index
-            type: pim_datasource_default
-            entity: '%pim_catalog.entity.group.class%'
-            repository_method: createDatagridQueryBuilder
+            type: pim_datasource_group
         columns:
             code:
                 label: Code
