+++ src/Pim/Bundle/EnrichBundle/Resources/config/datagrid/group_type.yml
@@ -1,21 +1,55 @@
 datagrid:
     group-type-grid:
-        extend: translatable-entity-grid
         options:
+            locale_parameter: localeCode
             entityHint: group type
         source:
             acl_resource: pim_enrich_grouptype_index
-            entity: '%pim_catalog.entity.group_type.class%'
+            type: pim_datasource_group_type
+        columns:
+            code:
+                label: Code
+            label:
+                label: Label
         properties:
+            id: ~
             edit_link:
+                type: url
                 route: pim_enrich_grouptype_edit
+                params:
+                    - id
             delete_link:
+                type: url
                 route: pim_enrich_grouptype_remove
+                params:
+                    - id
+        actions:
+            edit:
+                type:      navigate
+                label:     Edit
+                icon:      pencil
+                link:      edit_link
+                rowAction: true
+            delete:
+                type:  delete
+                label: Delete
+                icon:  trash
+                link:  delete_link
         sorters:
             columns:
                 code:
                     data_name: g.code
+                label:
+                    data_name: label
+            default:
+                code: '%oro_datagrid.extension.orm_sorter.class%::DIRECTION_ASC'
         filters:
             columns:
                 code:
+                    type:  string
+                    label: Code
                     data_name: g.code
+                label:
+                    type: string
+                    label: Label
+                    data_name: translation.label
