+++ src/Pim/Bundle/EnrichBundle/Resources/config/repositories.yml
@@ -9,9 +9,15 @@ parameters:
     pim_enrich.repository.category.class:         Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\CategoryRepository
     pim_enrich.repository.group.class:            Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\GroupRepository
     pim_enrich.repository.group_type.class:       Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\GroupTypeRepository
+    pim_enrich.repository.association_type.class: Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\AssociationTypeRepository
     pim_enrich.repository.attribute.class:        Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\AttributeRepository
     pim_enrich.repository.family.class:           Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\FamilyRepository
+    pim_enrich.repository.locale.class:           Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\LocaleRepository
+    pim_enrich.repository.channel.class:          Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\ChannelRepository
+    pim_enrich.repository.currency.class:         Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\CurrencyRepository
     pim_enrich.repository.attribute_group.class:  Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\AttributeGroupRepository
+    pim_enrich.repository.job_instance.class:     Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\JobInstanceRepository
+    pim_enrich.repository.job_execution.class:    Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository\JobExecutionRepository
 
 services:
     pim_enrich.repository.sequential_edit:
@@ -57,6 +63,12 @@ services:
             - '@doctrine.orm.entity_manager'
             - '%pim_catalog.entity.group_type.class%'
 
+    pim_enrich.repository.association_type:
+        class: '%pim_enrich.repository.association_type.class%'
+        arguments:
+            - '@doctrine.orm.entity_manager'
+            - '%pim_catalog.entity.association_type.class%'
+
     pim_enrich.repository.attribute:
         class: '%pim_enrich.repository.attribute.class%'
         arguments:
@@ -69,7 +81,25 @@ services:
         arguments:
             - '@pim_user.context.user'
             - '@doctrine.orm.entity_manager'
-            - '%pim_catalog.entity.attribute.class%'
+            - '%pim_catalog.entity.family.class%'
+
+    pim_enrich.repository.locale:
+        class: '%pim_enrich.repository.locale.class%'
+        arguments:
+            - '@doctrine.orm.entity_manager'
+            - '%pim_catalog.entity.locale.class%'
+
+    pim_enrich.repository.channel:
+        class: '%pim_enrich.repository.channel.class%'
+        arguments:
+            - '@doctrine.orm.entity_manager'
+            - '%pim_catalog.entity.channel.class%'
+
+    pim_enrich.repository.currency:
+        class: '%pim_enrich.repository.currency.class%'
+        arguments:
+            - '@doctrine.orm.entity_manager'
+            - '%pim_catalog.entity.currency.class%'
 
     pim_enrich.repository.attribute_group:
         class: '%pim_enrich.repository.attribute_group.class%'
@@ -77,6 +107,18 @@ services:
             - '@pim_user.context.user'
             - '@doctrine.orm.entity_manager'
             - '%pim_catalog.entity.attribute_group.class%'
+
+    pim_enrich.repository.job_instance:
+        class: '%pim_enrich.repository.job_instance.class%'
+        arguments:
+            - '@doctrine.orm.entity_manager'
+            - '%akeneo_batch.entity.job_instance.class%'
+
+    pim_enrich.repository.job_execution:
+        class: '%pim_enrich.repository.job_execution.class%'
+        arguments:
+            - '@doctrine.orm.entity_manager'
+            - '%akeneo_batch.entity.job_execution.class%'
 
     pim_enrich.repository.attribute_option.search:
         class: '%pim_enrich.repository.attribute_option.search.class%'
