+++ src/Pim/Bundle/ImportExportBundle/Entity/Repository/JobExecutionRepository.php
@@ -16,70 +16,6 @@ use Doctrine\ORM\EntityRepository;
 class JobExecutionRepository extends EntityRepository
 {
     /**
-     * Get data for the last operations widget
-     *
-     * @param array $types Job types to show
-     *
-     * @return array
-     */
-    public function getLastOperationsData(array $types)
-    {
-        $qb = $this->getLastOperationsQB($types);
-
-        return $qb->getQuery()->getArrayResult();
-    }
-
-    /**
-     * Get last operations query builder
-     *
-     * @param array $types
-     *
-     * @return \Doctrine\ORM\QueryBuilder
-     */
-    protected function getLastOperationsQB(array $types)
-    {
-        $qb = $this->createQueryBuilder('e');
-        $qb
-            ->select('e.id, e.startTime as date, j.type, j.label, e.status')
-            ->innerJoin('e.jobInstance', 'j')
-            ->orderBy('e.startTime', 'DESC')
-            ->setMaxResults(10);
-
-        if (!empty($types)) {
-            $qb->andWhere($qb->expr()->in('j.type', $types));
-        }
-
-        return $qb;
-    }
-
-    /**
-     * Create datagrid query builder
-     *
-     * @return \Doctrine\ORM\QueryBuilder
-     */
-    public function createDatagridQueryBuilder()
-    {
-        $qb = $this->createQueryBuilder('e');
-        $qb
-            ->addSelect('e.id')
-            ->addSelect('e.status AS status')
-            ->addSelect(
-                "CONCAT('pim_import_export.batch_status.', e.status) as statusLabel"
-            )
-            ->addSelect('e.startTime as date')
-            ->addSelect('j.code AS jobCode')
-            ->addSelect('j.label AS jobLabel')
-            ->addSelect('j.jobName as jobName')
-        ;
-
-        $qb->innerJoin('e.jobInstance', 'j');
-
-        $qb->andWhere('j.type = :jobType');
-
-        return $qb;
-    }
-
-    /**
      * Create job tracker datagrid query builder
      *
      * @return \Doctrine\ORM\QueryBuilder
