+++ src/Pim/Bundle/ImportExportBundle/Resources/config/datagrid/job_execution.yml
@@ -1,9 +1,7 @@
 datagrid:
     job-execution-grid:
         source:
-            type: pim_datasource_default
-            entity: '%akeneo_batch.entity.job_execution.class%'
-            repository_method: createDatagridQueryBuilder
+            type: pim_datasource_job_execution
         columns:
             code:
                 label:     Code
