+++ src/Pim/Bundle/ImportExportBundle/Resources/config/datagrid/job_profile.yml
@@ -1,9 +1,7 @@
 datagrid:
     job-profile-grid:
         source:
-            type: pim_datasource_default
-            entity: '%akeneo_batch.entity.job_instance.class%'
-            repository_method: createDatagridQueryBuilder
+            type: pim_datasource_job_instance
         columns:
             code:
                 label: Code
