+++ src/Pim/Bundle/ImportExportBundle/Resources/config/managers.yml
@@ -5,5 +5,5 @@ services:
     pim_import_export.manager.job_execution:
         class: '%pim_import_export.manager.job_execution.class%'
         arguments:
-            - '@pim_import_export.repository.job_execution'
+            - '@pim_enrich.repository.job_execution'
             - '@oro_security.security_facade'
