+++ src/Pim/Component/Catalog/Repository/ChannelRepositoryInterface.php
@@ -24,13 +24,6 @@ interface ChannelRepositoryInterface extends IdentifiableObjectRepositoryInterfa
     public function countAll();
 
     /**
-     * TODO: should be extracted in an enrich bundle repository
-     *
-     * @return mixed a query builder
-     */
-    public function createDatagridQueryBuilder();
-
-    /**
      * Get the deleted locales of a channel (the channel is updated but not flushed yet).
      *
      * @param ChannelInterface $channel
